+++ src/x3270/telnet.c
@@ -1476,7 +1476,7 @@ tn3270e_request(void)
 
 	trace_dsn("SENT %s %s DEVICE-TYPE REQUEST %.*s%s%s "
 		   "%s\n",
-	    cmd(SB), opt(TELOPT_TN3270E), strlen(termtype), tt_out + 5,
+	    cmd(SB), opt(TELOPT_TN3270E), (int)strlen(termtype), tt_out + 5,
 	    (try_lu != CN && *try_lu) ? " CONNECT " : "",
 	    (try_lu != CN && *try_lu) ? try_lu : "",
 	    cmd(SE));
